feat(docs): support granular API placeholders
