+++ scripts/generate-docs.ts
@@ -2,8 +2,6 @@ import { access, mkdir, readFile, writeFile } from 'node:fs/promises';
 import { join, relative, resolve } from 'node:path';
 import fm from 'front-matter';
 import markdownToHtml from 'zenn-markdown-html';
-import { formatDescription, formatType } from '@capacitor/docgen/dist/formatting';
-import { MarkdownTable } from '@capacitor/docgen/dist/markdown';
 import { JSDOM } from 'jsdom';
 import {
   type Locale,
@@ -36,73 +34,10 @@ import {
 } from './package-markdown';
 import { assertValidContentUpdatedAt, formatSitemapLastmod } from './seo-dates';
 import { loadRelatedArticlesByLibrary, type RelatedArticle } from './article-relations';
+import { apiMarkdown } from './docgen-api';
 
 const root = resolve(process.cwd());
 const docsRepositoryUrl = 'https://github.com/rdlabo-dev/website';
-
-function stripHtml(value: string): string {
-  const fragment = JSDOM.fragment(value);
-  for (const code of fragment.querySelectorAll('code')) {
-    code.replaceWith(`\`${code.textContent ?? ''}\``);
-  }
-  return fragment.textContent ?? '';
-}
-const tagText = (tags: any[], name: string) => tags?.find((tag) => tag.name === name)?.text ?? '';
-
-function apiMarkdown(source: any): Map<string, string> {
-  const entries = new Map<string, string>();
-  for (const method of source.api?.methods ?? []) {
-    const signature =
-      method.name === 'addListener' && method.parameters?.length
-        ? `addListener(${String(method.parameters[0].type).replace(/"/g, "'")}, ...)`
-        : `${method.name}(${method.parameters?.length ? '...' : ''})`;
-    const markdown = `#### \`method\` ${signature}\n${formatDescription(source, method.docs) || ''}\n\n\`${method.name}${method.signature}\`\n`;
-    const existing = entries.get(method.name);
-    entries.set(method.name, existing ? `${existing}\n${markdown}` : markdown);
-  }
-  for (const item of source.interfaces ?? []) {
-    const table = new MarkdownTable();
-    table.addHeader(['Prop', 'Type', 'Description', 'Default', 'Since']);
-    for (const property of item.properties ?? []) {
-      table.addRow([
-        `**\`${property.name}\`**`,
-        formatType(source, property.type).formatted,
-        formatDescription(source, property.docs),
-        tagText(property.tags, 'default'),
-        tagText(property.tags, 'since'),
-      ]);
-    }
-    table.removeEmptyColumns();
-    entries.set(
-      item.name,
-      `#### \`interface\` ${item.name}\n${formatDescription(source, item.docs) || ''}\n${stripHtml(table.toMarkdown().join('\n'))}\n`,
-    );
-  }
-  for (const item of source.typeAliases ?? []) {
-    const types = item.types
-      .map((type: any) => formatType(source, type.text).formatted)
-      .join(' | ');
-    entries.set(item.name, `#### \`type alias\` ${item.name}\n${stripHtml(types)}\n`);
-  }
-  for (const item of source.enums ?? []) {
-    const table = new MarkdownTable();
-    table.addHeader(['Member', 'Value', 'Description', 'Since']);
-    for (const member of item.members ?? []) {
-      table.addRow([
-        `**\`${member.name}\`**`,
-        formatType(source, member.value).formatted,
-        formatDescription(source, member.docs),
-        tagText(member.tags, 'since'),
-      ]);
-    }
-    table.removeEmptyColumns();
-    entries.set(
-      item.name,
-      `#### \`enum\` ${item.name}\n${stripHtml(table.toMarkdown().join('\n'))}\n`,
-    );
-  }
-  return entries;
-}
 
 async function renderCode(markdown: string): Promise<{ file: string; lines: string[] }> {
   const parsed = fm<{ title?: string; file?: string }>(markdown);
